requirements.txt
python-dotenv
numpy 
pandas
pymongo
certifi
pymongo[srv]
scikit-learn
pyaml
## -e .